+++ test/Operators/finitedifference/inference_examples.jl
@@ -189,7 +189,7 @@ function alloc_test_nested_expressions_1(cfield, ffield)
     (; cx, cy, cz, cϕ, cψ) = cfield
     ∇c = Operators.DivergenceF2C()
     wvec = Geometry.WVector
-    LB = Operators.LeftBiasedC2F(; bottom = Operators.SetValue(1))
+    LB = Operators.BottomBiasedC2F(; bottom = Operators.SetValue(1))
     @. cz = cx * cy * ∇c(wvec(LB(cy))) * ∇c(wvec(LB(cx))) * cϕ * cψ # Compile first
     p = @allocated begin
         @. cz = cx * cy * ∇c(wvec(LB(cy))) * ∇c(wvec(LB(cx))) * cϕ * cψ
@@ -202,7 +202,7 @@ function alloc_test_nested_expressions_2(cfield, ffield)
     (; cx, cy, cz, cϕ, cψ) = cfield
     ∇c = Operators.DivergenceF2C()
     wvec = Geometry.WVector
-    RB = Operators.RightBiasedC2F(; top = Operators.SetValue(1))
+    RB = Operators.TopBiasedC2F(; top = Operators.SetValue(1))
     @. cz = cx * cy * ∇c(wvec(RB(cy))) * ∇c(wvec(RB(cx))) * cϕ * cψ # Compile first
     p = @allocated begin
         @. cz = cx * cy * ∇c(wvec(RB(cy))) * ∇c(wvec(RB(cx))) * cϕ * cψ
@@ -216,7 +216,7 @@ function alloc_test_nested_expressions_3(cfield, ffield)
     Ic = Operators.InterpolateF2C()
     ∇c = Operators.DivergenceF2C()
     wvec = Geometry.WVector
-    LB = Operators.LeftBiasedC2F(; bottom = Operators.SetValue(1))
+    LB = Operators.BottomBiasedC2F(; bottom = Operators.SetValue(1))
     #! format: off
     @. cz = cx * cy * ∇c(wvec(LB(Ic(fy) * cx))) * ∇c(wvec(LB(Ic(fy) * cx))) * cϕ * cψ # Compile first
     p = @allocated begin
@@ -238,7 +238,7 @@ function alloc_test_nested_expressions_4(cfield, ffield)
         bottom = Operators.SetDivergence(0),
         top = Operators.SetDivergence(0),
     )
-    LB = Operators.LeftBiasedF2C(; bottom = Operators.SetValue(1))
+    LB = Operators.BottomBiasedF2C(; bottom = Operators.SetValue(1))
     #! format: off
     @. fz = fx * fy * ∇f(wvec(LB(If(cy) * fx))) * ∇f(wvec(LB(If(cy) * fx))) * fϕ * fψ # Compile first
     p = @allocated begin
@@ -412,7 +412,7 @@ function alloc_test_nested_expressions_13(
         bottom = Operators.SetValue(wvec(FT(0))),
         top = Operators.SetValue(wvec(FT(0))),
     )
-    LBC = Operators.LeftBiasedF2C(; bottom = Operators.SetValue(FT(0)))
+    LBC = Operators.BottomBiasedF2C(; bottom = Operators.SetValue(FT(0)))
     zero_bcs =
         (; bottom = Operators.SetValue(FT(0)), top = Operators.SetValue(FT(0)))
     I0f = Operators.InterpolateC2F(; zero_bcs...)
@@ -511,12 +511,12 @@ end
     alloc_test_c2f_interp(
         cfield,
         ffield,
-        Operators.LeftBiasedC2F(; bottom = Operators.SetValue(0)),
+        Operators.BottomBiasedC2F(; bottom = Operators.SetValue(0)),
     )
     alloc_test_c2f_interp(
         cfield,
         ffield,
-        Operators.RightBiasedC2F(; top = Operators.SetValue(0)),
+        Operators.TopBiasedC2F(; top = Operators.SetValue(0)),
     )
 
     alloc_test_derivative(