+++ test/Operators/spectralelement/conv_sphere_diffusion_vec.jl
@@ -19,8 +19,8 @@ include("sphere_sphericalharmonics.jl")
 function ∇²(u)
     scurl = Operators.Curl()
     sdiv = Operators.Divergence()
-    wcurl = Operators.WeakCurl()
-    wgrad = Operators.WeakGradient()
+    wcurl = Operators.Curl{Operators.WeakForm}()
+    wgrad = Operators.Gradient{Operators.WeakForm}()
 
     χ = Spaces.weighted_dss!(
         @. wgrad(sdiv(u)) - Geometry.Covariant12Vector(